backend/requirements.txt
Django>=4.2,<6.1
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.0
django-jazzmin>=3.0
Pillow>=10.0
gunicorn>=21.2
psycopg2-binary>=2.9
